+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Unreleased
+
+- Rolls originating from actors now automatically append the actor's Morale value to the end of the dice formula and display it in chat (e.g. `(2d6khx) + 2`). If Morale is 0 no change is made.
+
 ## 1.2.0
 
 - Add support for Foundry v10